Code refactor

+++ app/lib/features/user/bloc/user_config/user_config_bloc.dart
@@ -1,3 +1,4 @@
+import 'package:discord/discord.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:snitch/features/user/model/user_config_model.dart';
@@ -10,15 +11,17 @@ class UserConfigBloc extends Bloc<UserConfigEvent, UserConfigState> {
 
 
   UserConfigBloc({ required this.repository }) : super(UserConfigInitial()) {
-    on<UserConfigReadEvent>(_readConfig);
-    on<UserConfigCreateEvent>(_createConfig);
+    on<UserConfigReadEvent>(_read);
+    on<UserConfigCreateEvent>(_create);
+    on<UserConfigUpdateEvent>(_update);
   }
 
 
   final UserLocalRepository repository;
+  final discordValidator = const DiscordValidator();
 
 
-  Future<void> _readConfig(UserConfigReadEvent event, emit) async {
+  Future<void> _read(UserConfigReadEvent event, emit) async {
 
     try {
 
@@ -39,25 +42,57 @@ class UserConfigBloc extends Bloc<UserConfigEvent, UserConfigState> {
   }
 
 
-  Future<void> _createConfig(UserConfigCreateEvent event, emit) async {
+  Future<void> _create(UserConfigCreateEvent event, emit) async {
 
       try {
         emit(const UserConfigCreating());
+
+        final config = event.config;
+
+        final result = await discordValidator.check(config.token, config.guildId);
+
+        if (result is! DiscordValidatorStatusSuccess) {
+          throw result.message;
+        }
 
         await repository.deleteAll();
 
         final configs = await repository.create(event.config);
 
         if (configs == null) {
-          emit(const UserConfigEmpty());
+          throw 'Error while creating config';
         }
         else {
-          emit(UserConfigReaded(configs));
+          emit(UserConfigCreated(configs));
         }
 
       }
       catch (e) {
-        emit(UserConfigError(e.toString()));
+        emit(UserConfigCreateError(e.toString()));
+      }
+  }
+
+
+  Future<void> _update(UserConfigUpdateEvent event, emit) async {
+
+      try {
+        emit(const UserConfigUpdating());
+
+        final config = event.config;
+
+        final result = await discordValidator.check(config.token, config.guildId);
+
+        if (result is! DiscordValidatorStatusSuccess) {
+          throw result.message;
+        }
+
+        final configs = await repository.update(event.config);
+
+        emit(UserConfigUpdated(configs));
+
+      }
+      catch (e) {
+        emit(UserConfigUpdateError(e.toString()));
       }
   }
